+++ common/RELEASES.md
@@ -591,3 +591,13 @@ Based on:
 - Speakeasy CLI 1.59.0 (2.65.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [PyPI v0.33.1] https://pypi.org/project/codat-common/0.33.1 - common
+
+## 2023-08-17 00:15:13
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Common.yaml
+- Speakeasy CLI 1.73.1 (2.84.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.34.0] common
+### Releases
+- [PyPI v0.34.0] https://pypi.org/project/codat-common/0.34.0 - common